restaurantmanagement/src/main/java/com/raghu/restaurantmanagement/controller/CategoryController.java
/**
 * 
 */
package com.raghu.restaurantmanagement.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import com.raghu.restaurantmanagement.domain.Category;
import com.raghu.restaurantmanagement.service.CategoryService;

@RestController
@RequestMapping("/category")
public class CategoryController {

	@Autowired
	private CategoryService categoryService;

	public void setCategoryService(CategoryService categoryService) {
		this.categoryService = categoryService;
	}

	@RequestMapping(method = RequestMethod.POST)
	public ResponseEntity<Category> createCategory(@RequestBody Category category) {
		category = categoryService.save(category);
		return new ResponseEntity<Category>(HttpStatus.CREATED);
	}

	@RequestMapping(value = "/{categoryId}", method = RequestMethod.GET)
	public ResponseEntity<Category> getCategory(@PathVariable String categoryId) {

		Category category = categoryService.get(categoryId);
		if (category == null) {
			return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
		}
		return new ResponseEntity<Category>(category, HttpStatus.OK);
	}

}
